+++ resources/views/species/edit.blade.php
@@ -646,6 +646,27 @@
                 </div>
 
                 <div class="card mb-3">
+                    <div class="card-header">Others</div>
+                    <div class="card-body">
+                        <div class="form-group row">
+                            <label for="interesting_cultivar" class="col-md-3 col-form-label text-md-right">{{ __('Interesting cultivars') }} *</label>
+
+                            <div class="col-md-6">
+                                <input id="interesting_cultivar" type="text" class="form-control @error('interesting_cultivar') is-invalid @enderror" name="interesting_cultivar" value="{{ implode(',', $species->interesting_cultivar) }}" autofocus>
+
+                                <span class="form-text">Separate each cultivar with comas,<br>ex: <code>a cool cultivar,cultivar1,another cultivar</code></span>
+
+                                @error('interesting_cultivar')
+                                    <span class="invalid-feedback" role="alert">
+                                        <strong>{{ $message }}</strong>
+                                    </span>
+                                @enderror
+                            </div>
+                        </div>
+                    </div>
+                </div>
+
+                <div class="card mb-3">
                     <button type="submit" class="btn btn-lg btn-primary">
                         {{ __('Update') }}
                     </button>